Replace `src/card` with `src/ui/card`
- Update import paths
- Keep old implementation temporarily in `src/card-old`
