+++ packages/components/src/card-old/stories/_index.js
@@ -12,7 +12,7 @@ import CardFooter from '../footer';
 import CardHeader from '../header';
 import { getCardStoryProps } from './_utils';
 
-export default { title: 'Components/Card', component: Card };
+export default { title: 'Components/CardOld', component: Card };
 
 export const _default = () => {
 	const props = getCardStoryProps();
