+++ packages/components/src/card-old/stories/body.js
@@ -10,7 +10,7 @@ import Card from '../index';
 import CardBody from '../body';
 import { getCardStoryProps } from './_utils';
 
-export default { title: 'Components/Card/Body', component: CardBody };
+export default { title: 'Components/CardOld/Body', component: CardBody };
 
 export const _default = () => {
 	const props = getCardStoryProps();
