+++ packages/components/src/card-old/stories/divider.js
@@ -6,7 +6,7 @@ import CardBody from '../body';
 import CardDivider from '../divider';
 import { getCardStoryProps } from './_utils';
 
-export default { title: 'Components/Card/Divider', component: CardDivider };
+export default { title: 'Components/CardOld/Divider', component: CardDivider };
 
 export const _default = () => {
 	const props = getCardStoryProps();
