+++ packages/components/src/card-old/stories/footer.js
@@ -7,38 +7,38 @@ import { boolean, text } from '@storybook/addon-knobs';
  * Internal dependencies
  */
 import Card from '../index';
-import CardHeader from '../header';
+import CardFooter from '../footer';
 import Button from '../../button';
 import { FlexBlock, FlexItem } from '../../flex';
 import { getCardStoryProps } from './_utils';
 
-export default { title: 'Components/Card/Header', component: CardHeader };
+export default { title: 'Components/CardOld/Footer', component: CardFooter };
 
 export const _default = () => {
 	const props = getCardStoryProps();
-	const content = text( 'Header: children', 'Content' );
-	const isShady = boolean( 'Header: isShady', false );
+	const content = text( 'Footer: children', 'Content' );
+	const isShady = boolean( 'Footer: isShady', false );
 
 	return (
 		<Card { ...props }>
-			<CardHeader isShady={ isShady }>{ content }</CardHeader>
+			<CardFooter isShady={ isShady }>{ content }</CardFooter>
 		</Card>
 	);
 };
 
 export const alignment = () => {
 	const props = getCardStoryProps();
-	const content = text( 'Header: children', 'Content' );
-	const isShady = boolean( 'Header: isShady', false );
+	const content = text( 'Footer: children', 'Content' );
+	const isShady = boolean( 'Footer: isShady', false );
 
 	return (
 		<Card { ...props }>
-			<CardHeader isShady={ isShady }>
+			<CardFooter isShady={ isShady }>
 				<FlexBlock>{ content }</FlexBlock>
 				<FlexItem>
 					<Button variant="primary">Action</Button>
 				</FlexItem>
-			</CardHeader>
+			</CardFooter>
 		</Card>
 	);
 };
