+++ packages/components/src/card-old/stories/media.js
@@ -14,7 +14,7 @@ import CardHeader from '../header';
 import CardMedia from '../media';
 import { getCardStoryProps } from './_utils';
 
-export default { title: 'Components/Card/Media', component: CardMedia };
+export default { title: 'Components/CardOld/Media', component: CardMedia };
 
 const DummyImage = () => (
 	<img
