+++ packages/components/src/card/component.js
@@ -11,15 +11,15 @@ import { useMemo } from '@wordpress/element';
 /**
  * Internal dependencies
  */
-import { contextConnect } from '../context';
-import { Elevation } from '../../elevation';
-import { View } from '../../view';
+import { contextConnect } from '../ui/context';
+import { Elevation } from '../elevation';
+import { View } from '../view';
 import * as styles from './styles';
 import { useCard } from './hook';
-import CONFIG from '../../utils/config-values';
+import CONFIG from '../utils/config-values';
 
 /**
- * @param {import('../context').PolymorphicComponentProps<import('./types').CardProps, 'div'>} props
+ * @param {import('../ui/context').PolymorphicComponentProps<import('./types').CardProps, 'div'>} props
  * @param {import('react').Ref<any>} forwardedRef
  */
 function Card( props, forwardedRef ) {
