+++ packages/components/src/card/footer.js
@@ -1,34 +1,66 @@
 /**
  * External dependencies
  */
-import classnames from 'classnames';
+import { cx } from 'emotion';
+
+/**
+ * WordPress dependencies
+ */
+import { useMemo } from '@wordpress/element';
 
 /**
  * Internal dependencies
  */
-import { FooterUI } from './styles/card-styles';
-import { useCardContext } from './context';
+import { contextConnect, useContextSystem } from '../ui/context';
+import { Flex } from '../flex';
+import * as styles from './styles';
 
-export const defaultProps = {
-	isBorderless: false,
-	isShady: false,
-	size: 'medium',
-};
+/**
+ * @param {import('../ui/context').PolymorphicComponentProps<import('./types').CardFooterProps, 'div'>} props
+ * @param {import('react').Ref<any>} forwardedRef
+ */
+function CardFooter( props, forwardedRef ) {
+	const {
+		className,
+		justify = 'flex-end',
+		size = 'medium',
+		...otherProps
+	} = useContextSystem( props, 'CardFooter' );
 
-export function CardFooter( props ) {
-	const { className, isShady, ...additionalProps } = props;
-	const mergedProps = { ...defaultProps, ...useCardContext(), ...props };
-	const { isBorderless, size } = mergedProps;
-
-	const classes = classnames(
-		'components-card__footer',
-		isBorderless && 'is-borderless',
-		isShady && 'is-shady',
-		size && `is-size-${ size }`,
-		className
+	const classes = useMemo(
+		() =>
+			cx(
+				styles.borderRadius,
+				styles.headerFooter,
+				styles[ size ],
+				className
+			),
+		[ className, size ]
 	);
 
-	return <FooterUI { ...additionalProps } className={ classes } />;
+	return (
+		<Flex
+			{ ...otherProps }
+			className={ classes }
+			justify={ justify }
+			ref={ forwardedRef }
+		/>
+	);
 }
 
-export default CardFooter;
+/**
+ * `CardFooter` is a layout component, rendering the footer content of a `Card`.
+ *
+ * @example
+ * ```jsx
+ * import { Card, CardBody, CardFooter } from `@wordpress/components/ui`;
+ *
+ * <Card>
+ * 	<CardBody>...</CardBody>
+ * 	<CardFooter>...</CardFooter>
+ * </Card>
+ * ```
+ */
+const ConnectedCardFooter = contextConnect( CardFooter, 'CardFooter' );
+
+export default ConnectedCardFooter;
