+++ packages/components/src/card/header.js
@@ -1,34 +1,60 @@
 /**
  * External dependencies
  */
-import classnames from 'classnames';
+import { cx } from 'emotion';
+
+/**
+ * WordPress dependencies
+ */
+import { useMemo } from '@wordpress/element';
 
 /**
  * Internal dependencies
  */
-import { HeaderUI } from './styles/card-styles';
-import { useCardContext } from './context';
+import { contextConnect, useContextSystem } from '../ui/context';
+import { Flex } from '../flex';
+import * as styles from './styles';
 
-export const defaultProps = {
-	isBorderless: false,
-	isShady: false,
-	size: 'medium',
-};
-
-export function CardHeader( props ) {
-	const { className, isShady, ...additionalProps } = props;
-	const mergedProps = { ...defaultProps, ...useCardContext(), ...props };
-	const { isBorderless, size } = mergedProps;
-
-	const classes = classnames(
-		'components-card__header',
-		isBorderless && 'is-borderless',
-		isShady && 'is-shady',
-		size && `is-size-${ size }`,
-		className
+/**
+ * @param {import('../ui/context').PolymorphicComponentProps<import('./types').CardHeaderProps, 'div'>} props
+ * @param {import('react').Ref<any>} forwardedRef
+ */
+function CardHeader( props, forwardedRef ) {
+	const { className, size = 'medium', ...otherProps } = useContextSystem(
+		props,
+		'CardHeader'
 	);
 
-	return <HeaderUI { ...additionalProps } className={ classes } />;
+	const classes = useMemo(
+		() =>
+			cx(
+				styles.Header,
+				styles.borderRadius,
+				styles.headerFooter,
+				styles[ size ],
+				className
+			),
+		[ className, size ]
+	);
+
+	return (
+		<Flex { ...otherProps } className={ classes } ref={ forwardedRef } />
+	);
 }
 
-export default CardHeader;
+/**
+ * `CardHeader` is a layout component, rendering the header contents of a `Card`.
+ *
+ * @example
+ * ```jsx
+ * import { Card, CardBody, CardHeader } from `@wordpress/components/ui`;
+ *
+ * <Card>
+ * 	<CardHeader>...</CardHeader>
+ * 	<CardBody>...</CardBody>
+ * </Card>
+ * ```
+ */
+const ConnectedCardHeader = contextConnect( CardHeader, 'CardHeader' );
+
+export default ConnectedCardHeader;
