+++ packages/components/src/card/index.js
@@ -1,51 +1,7 @@
-/**
- * External dependencies
- */
-import classnames from 'classnames';
+export { default as Card } from './component';
+export { default as CardBody } from './body';
+export { default as CardFooter } from './footer';
+export { default as CardHeader } from './header';
+export { default as CardInnerBody } from './inner-body';
 
-/**
- * Internal dependencies
- */
-import { CardContext } from './context';
-import { CardUI } from './styles/card-styles';
-
-export const defaultProps = {
-	isBorderless: false,
-	isElevated: false,
-	size: 'medium',
-};
-
-export function Card( props ) {
-	const {
-		className,
-		isBorderless,
-		isElevated,
-		size,
-		...additionalProps
-	} = props;
-	const { Provider } = CardContext;
-
-	const contextProps = {
-		isBorderless,
-		isElevated,
-		size,
-	};
-
-	const classes = classnames(
-		'components-card',
-		isBorderless && 'is-borderless',
-		isElevated && 'is-elevated',
-		size && `is-size-${ size }`,
-		className
-	);
-
-	return (
-		<Provider value={ contextProps }>
-			<CardUI { ...additionalProps } className={ classes } />
-		</Provider>
-	);
-}
-
-Card.defaultProps = defaultProps;
-
-export default Card;
+export * from './hook';
