+++ packages/components/src/card/stories/index.js
@@ -1,17 +1,17 @@
 /**
  * Internal dependencies
  */
-import { Grid } from '../../../grid';
-import { Text } from '../../../text';
-import { VStack } from '../../../v-stack';
+import { Grid } from '../../grid';
+import { Text } from '../../text';
+import { VStack } from '../../v-stack';
 import { Card, CardBody, CardFooter } from '../index';
-import { Divider } from '../../../divider';
-import { Heading } from '../../../heading';
-import Button from '../../../button';
+import { Divider } from '../../divider';
+import { Heading } from '../../heading';
+import Button from '../../button';
 
 export default {
 	component: Card,
-	title: 'G2 Components (Experimental)/Card',
+	title: 'Components/Card',
 };
 
 const ExampleCard = () => {
