--- packages/components/src/ui/card/index.js
@@ -1,7 +1,0 @@
-export { default as Card } from './component';
-export { default as CardBody } from './body';
-export { default as CardFooter } from './footer';
-export { default as CardHeader } from './header';
-export { default as CardInnerBody } from './inner-body';
-
-export * from './hook';
