+++ packages/components/src/ui/index.js
@@ -1,4 +1,3 @@
-export * from './card';
 export * from './control-group';
 export * from './control-label';
 export * from './form-group';
