+++ packages/components/src/ui/popover/styles.js
@@ -6,7 +6,7 @@ import { css } from 'emotion';
 /**
  * Internal dependencies
  */
-import { CardBody } from '../card';
+import { CardBody } from '../../card';
 import * as ZIndex from '../../utils/z-index';
 import CONFIG from '../../utils/config-values';
